+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * (theimo1221) Fix edge case, where a manual timebased toggle would not turn on the device at it would be off by it's
   time settings
+* (theimo1221) Prevent AC-Device Param-NG in case of not yet calculated desired temperature
 
 ## 3.0.0-alpha.13 (2024-03-17)